Added an option to insert tracker with only the custom vars in ajax requests
The option is named "insert_custom_vars_if_ajax". When it's set, in ajax
requests, if custom vars have been defined, the tracker will be inserted
but will contain only the push for these vars, this is no a full re-declaration.
You still need to push the _trackPageView or _trackEvent with js, for
example.

This is primary designed like this to work with pjax, because
a) the custom vars may change over the pages loaded in pjax
b) pjax push the _trackPageView at the end of each request

This have been implemented only in the async tracker.